% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{do_seasonal_difference}
\alias{do_seasonal_difference}
\title{Do first-order seasonal differencing (go from original time series values to
seasonally differenced time series values).}
\usage{
do_seasonal_difference(y, ts_frequency)
}
\arguments{
\item{y}{a univariate time series or numeric vector.}

\item{ts_frequency}{frequency of time series.  Must be provided if y is not
of class "ts".  See the help for stats::ts for more.}
}
\value{
a seasonally differenced time series object (of class 'ts'),
  padded with leading NAs.
}
\description{
Do first-order seasonal differencing (go from original time series values to
seasonally differenced time series values).
}
