% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{fit_sarima}
\alias{fit_sarima}
\title{Estimate SARIMA model}
\usage{
fit_sarima(y, ts_frequency, transformation = "box-cox",
  seasonal_difference = TRUE, d = NA, D = NA)
}
\arguments{
\item{y}{a univariate time series or numeric vector.}

\item{ts_frequency}{frequency of time series.  Must be provided if y is not
of class "ts".  See the help for stats::ts for more.}

\item{transformation}{character specifying transformation type:
"box-cox", "log", "forecast-box-cox", or "none".  See details for more.}

\item{seasonal_difference}{boolean; take a seasonal difference before passing
to auto.arima?}

\item{d}{order of first differencing argument to auto.arima.}

\item{D}{order of seasonal differencing argument to auto.arima.}
}
\value{
a SARIMA model fit
}
\description{
Estimate SARIMA model
}
\details{
This function is a wrapper around forecast::auto.arima, providing
  some useful defaults for preliminary transformations of the data.
  Formal and informal experimentation has shown these preliminary
  transformations to be helpful with a few infectious disease time series
  data sets.  Note that if any transformation was specified or the
  seasonal_difference argument was TRUE in the call to this function, only
  prediction/forecast utilities provided by the sarimaTD package can be
  used!  We have found that using the default arguments for transformation,
  seasonal_difference, d, and D, yields good performance.
}
