% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{interpolate_and_clean_missing}
\alias{interpolate_and_clean_missing}
\title{Remove leading values that are infinite or missing, and replace all internal
sequences of infinite or missing values by linearly interpolating between
the surrounding observed (finite) values.  Used in prediction methods}
\usage{
interpolate_and_clean_missing(y)
}
\arguments{
\item{y}{a univariate time series or numeric vector}
}
\value{
a vector of the same class as y
}
\description{
Remove leading values that are infinite or missing, and replace all internal
sequences of infinite or missing values by linearly interpolating between
the surrounding observed (finite) values.  Used in prediction methods
}
