% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{invert_bc_transform}
\alias{invert_bc_transform}
\title{Invert a Box-Cox transformation.  See car::bcPower for the original
transformation.}
\usage{
invert_bc_transform(b, lambda, gamma)
}
\arguments{
\item{b}{a univariate numeric vector.}

\item{lambda}{exponent for Box-Cox transformation}

\item{gamma}{offset for Box-Cox transformation}
}
\value{
a transformed object of the same class as y
}
\description{
Invert a Box-Cox transformation.  See car::bcPower for the original
transformation.
}
