% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{invert_bcn_transform}
\alias{invert_bcn_transform}
\title{Invert a Box-Cox (with negatives allowed) transformation.  See car::bcnPower
for the original transformation.}
\usage{
invert_bcn_transform(b, lambda, gamma)
}
\arguments{
\item{b}{a univariate numeric vector.}

\item{lambda}{exponent for Box-Cox transformation}

\item{gamma}{offset for Box-Cox transformation}
}
\value{
a transformed object of the same class as y
}
\description{
Invert a Box-Cox (with negatives allowed) transformation.  See car::bcnPower
for the original transformation.
}
\details{
Undoing the Box-Cox step is straightforward, but I could not find a
  closed-form way to undo the second step (but I may just be missing it).
  Currently, just using optim to minimize square difference from z and
  transformed y, as a function of y.  There must be something faster?
}
