% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{invert_difference}
\alias{invert_difference}
\title{Invert first-order and seasonal differencing (go from seasonally differenced
time series to original time series).}
\usage{
invert_difference(dy, y, d, D, frequency)
}
\arguments{
\item{dy}{a first-order and/or seasonally differenced univariate time series
with values like y_{t} - y_{t - ts_frequency}}

\item{y}{a univariate time series or numeric vector with values like
y_{t - ts_frequency}.}

\item{d}{order of first differencing}

\item{D}{order of seasonal differencing}

\item{frequency}{frequency of time series.  Must be provided if y is not
of class "ts" and D > 0.  See the help for stats::ts for more.}
}
\value{
a time series object (of class 'ts')
}
\description{
Invert first-order and seasonal differencing (go from seasonally differenced
time series to original time series).
}
\details{
y may have longer length than dy.  It is assumed that dy "starts"
  one time index after y "ends": that is, if y is of length T, d = 0, and
  D = 1 then dy[1] = y[T + 1] - y[T + 1 - ts_frequency]
}
