% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{invert_initial_transform}
\alias{invert_initial_transform}
\title{Invert an initial transformation of time series data.}
\usage{
invert_initial_transform(y, transformation, bc_params)
}
\arguments{
\item{y}{a univariate time series or numeric vector.}

\item{transformation}{character specifying transformation type:
"box-cox", "log", "forecast-box-cox", or "none".}

\item{bc_params}{(required if transformation is "box-cox").  Parameters for
the Box-Cox transformation.  bc_params$lambda should be the result of a
call to car::powerTransform(y + bc_params$gamma, family = "bcPower")}
}
\value{
a transformed object of the same class as y
}
\description{
Invert an initial transformation of time series data.
}
