% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{invert_seasonal_difference}
\alias{invert_seasonal_difference}
\title{Invert first-order seasonal differencing (go from seasonally differenced time
series values to original time series values).}
\usage{
invert_seasonal_difference(dy, y, ts_frequency)
}
\arguments{
\item{dy}{a first-order seasonally differenced univariate time series with
values like y_{t} - y_{t - ts_frequency}}

\item{y}{a univariate time series or numeric vector with values like
y_{t - ts_frequency}.}

\item{ts_frequency}{frequency of time series.  Must be provided if y is not
of class "ts".  See the help for stats::ts for more.}
}
\value{
a time series object (of class 'ts')
}
\description{
Invert first-order seasonal differencing (go from seasonally differenced time
series values to original time series values).
}
\details{
y may have longer length than dy.  It is assumed that dy "starts"
  one time index after y "ends": that is, if y is of length T then
  dy[1] = y[T + 1] - y[T + 1 - ts_frequency]
}
