% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{sample_predictive_trajectories_arima}
\alias{sample_predictive_trajectories_arima}
\title{Simulate predictive trajectories from an ARIMA model}
\usage{
sample_predictive_trajectories_arima(
  object,
  h = ifelse(object$arma[5] > 1, 2 * object$arma[5], 10),
  bootstrap = FALSE,
  nsim = 1000,
  ...
)
}
\arguments{
\item{object}{an Arima fit object (with class "Arima")}

\item{h}{number of time steps forwards to simulate}

\item{nsim}{number of sample trajectories to simulate}
}
\value{
an nsim by h matrix with simulated values
}
\description{
This is directly taken from the forecast.Arima function from the forecast
package, but it returns the simulated trajectories which were not returned in
the original function definition.  This function calls
forecast:::simulate.Arima, which is NOT exported from the forecast package!!
}
