% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{sample_predictive_trajectories_arima_wrapper}
\alias{sample_predictive_trajectories_arima_wrapper}
\title{A wrapper around sample_predictive_trajectories_arima}
\usage{
sample_predictive_trajectories_arima_wrapper(sarima_fit, newdata, h = 1,
  npaths = 1)
}
\arguments{
\item{sarima_fit}{a sarima fit, probably as returned by fit_sarima}

\item{h}{number of time steps forwards to simulate}

\item{npaths}{number of sample trajectories to simulate}
}
\value{
an npaths by h matrix with simulated values
}
\description{
This function does a few things worth noting.  It linearly interpolates
missing values on the interior of the time series, which is necessary for
some model fits with moving average components.  It also handles
transformations and possible seasonal differencing that may have been done
in the fit_sarima function.
}
