% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarimaTD-package.R
\docType{package}
\name{sarimaTD}
\alias{sarimaTD}
\alias{sarimaTD-package}
\title{sarimaTD - SARIMA Models with Transformations and Seasonal Differencing}
\description{
SARIMA Models with Transformations and Seasonal Differencing
}
\author{
Evan L. Ray <eray@mtholyoke.edu>
}
