% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarimaTD_FF.R
\name{sarimaTD_FF}
\alias{sarimaTD_FF}
\title{sarimaTD interface via ForecastFramework}
\description{
Allows the user to use functionality provided by ForecastFramework with
sarimaTD models.
}
\section{Usage}{

\preformatted{sarimaTD_model = sarimaTD_FF$new()

sarimaTD_model$fit(data,
    period = 52,
    transformation = "box-cox",
    seasonal_difference = TRUE)

sarimaTD_model$forecast(newdata,
    steps = 1,
    nsim = 1000)

sarimaTD_model$data()

sarimaTD_model$frequency()

sarimaTD_model$nsim(nsim)

sarimaTD_model$models()
}
}

\section{Arguments}{

\code{data} A data set to fit sarimaTD models to, in a format compatible with
    ForecastFramework::IncidenceMatrix

\code{frequency} frequency of time series.  Must be provided if y is not
    of class "ts".  See the help for stats::ts for more.

\code{transformation} character specifying transformation type:
    "box-cox", "log", "forecast-box-cox", or "none".  See details of
    sarimaTD::fit_sarima for more.

\code{seasonal_difference} boolean; take a seasonal difference before passing
    to auto.arima?

\code{newdata} new data from which to predict forward, in a format compatible
    with ForecastFramework::IncidenceMatrix

\code{steps} If \code{TRUE}, force the Python process to terminate
  using a system call.

\code{nsim} Number of simulations to generate during forecasting
}

\section{Methods}{

\code{$new()} Initialize a sarimaTD model object.

\code{$fit()} Fit a set of models to the provided data, one model per
  location.

\code{$forecast()} Generate forecasts for each location.

\code{$data()} Get the data set used for the model fit.

\code{$nsim()} Get or set the number of simulations to generate during
    forecasting.

\code{$frequency()} Get the time series frequency of the data used for the
    model fit.

\code{$transformation()} Get the transformation performed before fitting the
    model.

\code{$seasonal_difference()} Get the indicator of whether or not first-order
    seasonal differencing is performed before fitting the model.

\code{$models()} Get a list of sarimaTD model fit objects, one for each
    location in the training data.
}

\examples{
vignette(package="sarimaTD", topic="sarimaTD_FF")
}
