% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarimautils-package.R
\docType{package}
\name{sarimautils}
\alias{sarimautils}
\alias{sarimautils-package}
\alias{sarimautils-package}
\title{sarimautils - SARIMA Model Utility Functions}
\description{
SARIMA Model Utility Functions
}
\author{
Evan L. Ray <eray@mtholyoke.edu>
}
