% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{simulate.sarimaTD}
\alias{simulate.sarimaTD}
\title{Simulate predictive trajectories from a SARIMA model with class sarima_td}
\usage{
\method{simulate}{sarimaTD}(object, nsim = 1, seed = NULL, newdata,
  h = 1)
}
\arguments{
\item{object}{a sarima fit of class "sarima_td", as returned by fit_sarima}

\item{nsim}{number of sample trajectories to simulate}

\item{seed}{either `NULL` or an integer that will be used in a call to
`set.seed` before simulating the response vectors.  If set, the value is
saved as the "seed" attribute of the returned value.  The default, `Null`,
will not change the random generator state, and return `.Random.seed`
as the "seed" attribute}

\item{newdata}{new data to simulate forward from}

\item{h}{number of time steps forwards to simulate}
}
\value{
an nsim by h matrix with simulated values
}
\description{
This function does a few things worth noting.  It linearly interpolates
missing values on the interior of the time series, which is necessary for
some model fits with moving average components.  It also handles
transformations and possible seasonal differencing that may have been done
in the fit_sarima function.
}
